+++ admin/includes/view_all_user.php
@@ -100,15 +100,27 @@ if (isset($_GET['change_to_sub'])) {
 
 
 if (isset($_GET['delete'])) {
-  
-  $the_user_id = $_GET['delete'];
 
-  $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-  $user_delete_query = mysqli_query($connection, $query);
+  if (isset($_SESSION['user_role'])) {
+    
+     if ($_SESSION['user_role'] == 'amdin') {
+        
 
-  confirm($user_delete_query);
+        $the_user_id =mysqli_real_escape_string($connection, $_GET['delete']);
 
-  header("Location: users.php");
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        $user_delete_query = mysqli_query($connection, $query);
+
+        confirm($user_delete_query);
+
+        header("Location: users.php");
+
+
+      } 
+
+
+  }
+
 }
